Fitter: Rename fitter module to base
