+++ Orange/modelling/base.py
